Reorganize the example a bit
